RealTimeControlSystem/requirements.txt
aqtinstall==2.0.5
beautifulsoup4==4.10.0
Brotli==1.0.9
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.10
cmake==3.22.1
cycler==0.11.0
fonttools==4.28.5
idna==3.3
importlib-metadata==4.10.0
importlib-resources==5.4.0
kiwisolver==1.3.2
matplotlib==3.5.1
multivolumefile==0.2.3
numpy==1.21.5
packaging==21.3
pandas==1.3.5
patch==1.16
Pillow==9.0.0
py7zr==0.17.2
pybcj==0.5.0
pycryptodomex==3.12.0
pygame==2.1.2
pygame-gui==0.6.0
pygame-menu==4.0.2
pylsl==1.15.0
pyparsing==3.0.6
pyperclip==1.8.2
pyppmd==0.17.3
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyqtgraph==0.12.3
pyserial==3.5
python-dateutil==2.8.2
python-i18n==0.3.9
pytz==2021.3
PyYAML==6.0
pyzstd==0.15.1
qtgui==0.0.1
requests==2.27.1
scipy==1.7.3
semantic-version==2.8.5
six==1.16.0
soupsieve==2.3.1
texttable==1.6.4
torch==1.10.1
typing_extensions==4.0.1
urllib3==1.26.8
zipp==3.7.0
